refactor(item_extractor): move treesitter-related logic to `ts.lua`
